Simplify math/countSteps. No more ternary and ~~
countSteps contained a ternary and a ~~ snake operator which was supposedly
fixing the poisonous NaN from calculating 0 mod overflow. NaN is only the
result of a mod when the right-hand side of the operator is 0. The ternary
already prevented that. The snake operator was also dangerous here because
it not only converts to Number, it converts to an integer! For step fractures
this would not work.

Amended the test spec to reflect this behaviour.