shorthand syntax to object/* & collection/* functions.
had to change the `object/every` structure to avoid circular dependencies and
on most collections methods it was just a matter of adding tests to confirm
behavior and avoid regressions in the future.

also *improved* `object/every` to make it consider any *falsy* value as
a "failure" so now behavior matches `array/every` and string shorthand syntax
can be used to check against any *falsy* values and property existence.

see #15