Add verification for arguments object
I believe most people passing an arguments object to `lang/toArray` just want
to get rid of the pseudo-array gotchas and convert it to an array.
To achieve that, using the Array.prototype.slice should be faster and save more
resources than iterating the object and create an array from scratch.